requirements_current.txt
about-time==4.2.1
absl-py==2.1.0
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiohttp-cors==0.7.0
aiosignal==1.3.2
alabaster==1.0.0
alembic==1.14.1
alive-progress==3.2.0
annotated-types==0.7.0
anyio==4.8.0
appnope==0.1.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
astunparse==1.6.3
async-lru==2.0.4
attrs==25.1.0
autograd==1.7.0
ax-platform==0.5.0
babel==2.17.0
bayesian-optimization==2.0.3
bcrypt==4.2.1
beautifulsoup4==4.13.3
black==25.1.0
bleach==6.2.0
blinker==1.9.0
botorch==0.13.0
cachetools==5.5.2
catboost==1.2.7
category_encoders==2.8.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==3.1.1
cma==3.2.2
colorama==0.4.6
colorful==0.5.6
colorlog==6.9.0
comm==0.2.2
contourpy==1.3.1
coverage==7.6.12
cryptography==44.0.1
cycler==0.12.1
dash==2.18.2
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
dask==2025.2.0
databricks-sdk==0.44.1
deap==1.4.2
debugpy==1.8.12
decorator==5.2.0
defusedxml==0.7.1
Deprecated==1.2.18
dill==0.3.9
distlib==0.3.9
distributed==2025.2.0
dnspython==2.7.0
docker==7.1.0
docker-pycreds==0.4.0
docutils==0.21.2
ecdsa==0.19.0
email_validator==2.2.0
execnet==2.1.1
executing==2.2.0
fastapi==0.115.8
fastjsonschema==2.21.1
filelock==3.16.1
flake8==7.1.2
Flask==3.0.3
flatbuffers==25.2.10
fonttools==4.56.0
fqdn==1.5.1
frozenlist==1.5.0
fsspec==2024.10.0
gast==0.6.0
gitdb==4.0.12
GitPython==3.1.44
google-api-core==2.24.1
google-auth==2.38.0
google-pasta==0.2.0
googleapis-common-protos==1.68.0
gpytorch==1.14
grapheme==0.6.0
graphene==3.4.3
graphql-core==3.2.6
graphql-relay==3.2.0
graphviz==0.20.3
grpcio==1.70.0
gunicorn==23.0.0
h11==0.14.0
h5py==3.13.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
imageio==2.37.0
imagesize==1.4.1
imbalanced-learn==0.13.0
importlib_metadata==8.5.0
iniconfig==2.0.0
ipykernel==6.29.5
ipython==8.32.0
ipywidgets==8.1.5
isoduration==20.11.0
itsdangerous==2.2.0
jaxtyping==0.2.38
jedi==0.19.2
Jinja2==3.1.4
joblib==1.4.2
json5==0.10.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.3.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
keras==3.8.0
kiwisolver==1.4.8
lazy_loader==0.4
libclang==18.1.1
lightgbm==4.6.0
lime==0.2.0.1
linear-operator==0.6
llvmlite==0.44.0
locket==1.0.0
Mako==1.3.9
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.10.0
matplotlib-inline==0.1.7
mccabe==0.7.0
-e git+https://github.com/bddupre92/mdt_Test.git@0bad148d81e7eaca92ebe84cf4d72b895ef4f830#egg=mdt_test
mdurl==0.1.2
memory-profiler==0.61.0
mistune==3.1.2
ml-dtypes==0.3.2
mlflow==2.20.2
mlflow-skinny==2.20.2
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
multipledispatch==1.0.0
mypy==1.15.0
mypy-extensions==1.0.0
namex==0.0.8
narwhals==1.27.1
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nbsphinx==0.9.6
nest-asyncio==1.6.0
networkx==3.4.2
notebook==7.3.2
notebook_shim==0.2.4
numba==0.61.0
numpy==1.26.4
opencensus==0.11.4
opencensus-context==0.1.3
opentelemetry-api==1.30.0
opentelemetry-sdk==1.30.0
opentelemetry-semantic-conventions==0.51b0
opt_einsum==3.4.0
optree==0.14.0
optuna==4.2.1
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
partd==1.4.2
passlib==1.7.4
pathspec==0.12.1
patsy==1.0.1
pdflatex==0.1.3
pexpect==4.9.0
pillow==11.0.0
platformdirs==4.3.6
plotly==6.0.0
pluggy==1.5.0
polars==1.23.0
prometheus_client==0.21.1
prompt_toolkit==3.0.50
propcache==0.3.0
proto-plus==1.26.0
protobuf==4.25.6
psutil==7.0.0
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
py-spy==0.4.0
pyarrow==18.1.0
pyasn1==0.4.8
pyasn1_modules==0.4.1
pycodestyle==2.12.1
pycparser==2.22
pydantic==2.10.6
pydantic-settings==2.8.0
pydantic_core==2.27.2
pyflakes==3.2.0
Pygments==2.19.1
PyJWT==2.10.1
pymoo==0.6.1.3
pyparsing==3.2.1
pyre-extensions==0.0.32
pyro-api==0.1.2
pyro-ppl==1.9.1
pytest==8.3.4
pytest-asyncio==0.25.3
pytest-cov==6.0.0
pytest-progress==1.3.0
pytest-sugar==1.0.0
pytest-xdist==3.6.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.4.0
python-json-logger==3.2.1
python-multipart==0.0.20
pytz==2025.1
PyYAML==6.0.2
pyzmq==26.2.1
ray==2.42.1
referencing==0.36.2
requests==2.32.3
retrying==1.3.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
roman-numerals-py==3.1.0
rpds-py==0.23.1
rsa==4.9
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
Send2Trash==1.8.3
sentry-sdk==2.22.0
setproctitle==1.3.5
setuptools==75.8.0
shap==0.46.0
six==1.17.0
sklearn-compat==0.1.3
slicer==0.0.8
smart-open==7.1.0
smmap==5.0.2
sniffio==1.3.1
snowballstemmer==2.2.0
sortedcontainers==2.4.0
soupsieve==2.6
Sphinx==8.2.1
sphinx-rtd-theme==3.0.2
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
SQLAlchemy==2.0.38
sqlparse==0.5.3
stack-data==0.6.3
starlette==0.45.3
statsmodels==0.14.4
sympy==1.13.3
tabulate==0.9.0
tblib==3.0.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
tensorflow==2.16.2
tensorflow-macos==2.16.2
termcolor==2.5.0
terminado==0.18.1
threadpoolctl==3.5.0
tifffile==2025.2.18
tinycss2==1.4.0
toolz==1.0.0
torch==2.7.0.dev20250223
torchaudio==2.6.0.dev20250223
torchvision==0.22.0.dev20250223
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
types-python-dateutil==2.9.0.20241206
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2025.1
uri-template==1.3.0
urllib3==2.3.0
uvicorn==0.34.0
virtualenv==20.29.2
wadler_lindig==0.1.3
wandb==0.19.7
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.0.6
wheel==0.45.1
widgetsnbextension==4.0.13
wrapt==1.17.2
xgboost==2.1.4
yarl==1.18.3
zict==3.0.0
zipp==3.21.0
